# Conflicts:
#	src/main/java/dat3/openai_demo/api/MealMentorController.java

+++ src/main/java/dat3/openai_demo/api/MealMentorController.java
@@ -23,17 +23,29 @@ public class MealMentorController {
   }
 
 
-  @PostMapping()
+
+    @PostMapping()
       public MyResponse generatePrompt(@RequestBody UserPromptResponse userPromptResponse){
 
-      String usernamePrompting = userPromptResponse.getUsername();
-      UserResponse user = userService.getUser(usernamePrompting);
+            // Check if the user has credits
+        if (userService.getUser(userPromptResponse.getUsername()).getCredits() <= 0) {
 
-      userService.removeOneCreditFromUser(usernamePrompting);
+            return new MyResponse("{\n" +
+                    "    \"Message\": \"USER HAS NO CREDITS\"\n" +
+                    "}");
 
+        } else {
 
-      System.out.println("****** " + user.getUsername() + " is PROMPTING *******\n" +
-              "****** and he/she has got " + userService.getUser(usernamePrompting).getCredits() + " credits left *****");
+            // Removing one credit from the user's credits
+            String usernamePrompting = userPromptResponse.getUsername();
+            UserResponse user = userService.getUser(usernamePrompting);
+            userService.removeOneCreditFromUser(usernamePrompting);
+
+            // Printing how many cr
+            System.out.println("****** " + user.getUsername() + " is PROMPTING *******\n" +
+                    "****** and he/she has got " +
+                    userService.getUser(usernamePrompting).getCredits()
+                    + " credits left *****");
 
       String userPrompt = "I am a " + user.getAge() + " old "
               + user.getSex() +
@@ -41,9 +53,12 @@ public class MealMentorController {
               "The recipes should include" + userPromptResponse.getMealChecklist() + "and must not include " + user.getAllergies()
               + ". My goals are" + user.getGoals()
 
-               + " i want the recipe made for " /* + userPromptResponse.getAmountOfDays() + "number of days" */ ;
+                    + " i want the recipe made for " /* + userPromptResponse.getAmountOfDays() + "number of days" */ ;
 
-      return openAiService.makeRequest(userPrompt, SYSTEM_MESSAGE);
+            return openAiService.makeRequest(userPrompt, SYSTEM_MESSAGE);
+        }
+
+
 
   }
